backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
requests
pandas
Faker
chromadb
sentence-transformers
groq
